+++ dwds/test/build_daemon_evaluate_test.dart
@@ -6,7 +6,6 @@
 @Timeout(Duration(minutes: 2))
 
 import 'package:test/test.dart';
-import 'package:test_common/test_sdk_configuration.dart';
 
 import 'evaluate_common.dart';
 import 'fixtures/context.dart';
@@ -16,13 +15,9 @@ void main() async {
   // Enable verbose logging for debugging.
   final debug = false;
 
-  final provider = TestSdkConfigurationProvider(verbose: debug);
-  tearDownAll(provider.dispose);
-
   for (var nullSafety in NullSafety.values) {
     group('${nullSafety.name} null safety |', () {
       testAll(
-        provider: provider,
         compilationMode: CompilationMode.buildDaemon,
         nullSafety: nullSafety,
         debug: debug,