+++ dwds/test/fixtures/context.dart
@@ -2,7 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
@@ -35,7 +34,8 @@ import 'package:test/test.dart';
 import 'package:test_common/logging.dart';
 import 'package:test_common/test_sdk_configuration.dart';
 import 'package:vm_service/vm_service.dart';
-import 'package:webdriver/async_io.dart';
+// ignore: deprecated_member_use
+import 'package:webdriver/io.dart';
 import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart';
 
 import 'project.dart';
@@ -55,7 +55,6 @@ enum CompilationMode { buildDaemon, frontendServer }
 class TestContext {
   final TestProject project;
   final NullSafety nullSafety;
-  final TestSdkConfigurationProvider sdkConfigurationProvider;
 
   /// Top level directory in which we run the test server, e.g.
   /// "/workstation/webdev/fixtures/_testSound".
@@ -144,8 +143,7 @@ class TestContext {
   /// External VM service.
   VmServiceInterface get vmService => debugConnection.vmService;
 
-  TestContext(this.project, this.sdkConfigurationProvider)
-      : nullSafety = project.nullSafety {
+  TestContext(this.project) : nullSafety = project.nullSafety {
     DartUri.currentDirectory = workingDirectory;
 
     project.validate();
@@ -176,13 +174,14 @@ class TestContext {
     bool isInternalBuild = false,
     List<String> experiments = const <String>[],
   }) async {
+    // Generate missing SDK assets if needed.
+    final sdkConfigurationProvider =
+        TestSdkConfigurationProvider(verboseCompiler: verboseCompiler);
     final sdkLayout = sdkConfigurationProvider.sdkLayout;
+    final configuration = await sdkConfigurationProvider.configuration;
+    configuration.validate();
 
     try {
-      // Make sure configuration was created correctly.
-      final configuration = await sdkConfigurationProvider.configuration;
-      configuration.validate();
-
       DartUri.currentDirectory = workingDirectory;
       configureLogWriter();
 
@@ -222,7 +221,7 @@ class TestContext {
             'Could not start ChromeDriver. Is it installed?\nError: $e');
       }
 
-      await Process.run(sdkLayout.dartPath, ['pub', 'upgrade'],
+      await Process.run(dartPath, ['pub', 'upgrade'],
           workingDirectory: workingDirectory);
 
       ExpressionCompiler? expressionCompiler;
@@ -244,8 +243,8 @@ class TestContext {
                 '--enable-experiment=$experiment',
               '--verbose',
             ];
-            _daemonClient = await connectClient(
-                sdkLayout.dartPath, workingDirectory, options, (log) {
+            _daemonClient =
+                await connectClient(workingDirectory, options, (log) {
               final record = log.toLogRecord();
               final name =
                   record.loggerName == '' ? '' : '${record.loggerName}: ';
@@ -390,7 +389,6 @@ class TestContext {
         ddcService,
         isFlutterApp,
         isInternalBuild,
-        sdkLayout,
       );
 
       _appUrl = basePath.isEmpty