+++ dwds/test/fixtures/server.dart
@@ -16,8 +16,9 @@ import 'package:dwds/src/services/expression_compiler_service.dart';
 import 'package:dwds/src/utilities/server.dart';
 import 'package:logging/logging.dart';
 import 'package:shelf/shelf.dart';
-import 'package:test_common/test_sdk_layout.dart';
 import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart';
+
+import 'utilities.dart';
 
 Logger _logger = Logger('TestServer');
 
@@ -81,7 +82,6 @@ class TestServer {
     ExpressionCompilerService? ddcService,
     bool isFlutterApp,
     bool isInternalBuild,
-    TestSdkLayout sdkLayout,
   ) async {
     var pipeline = const Pipeline();
 
@@ -116,13 +116,13 @@ class TestServer {
         urlEncoder: urlEncoder,
         expressionCompiler: expressionCompiler,
         isInternalBuild: isInternalBuild,
-        isFlutterApp: isFlutterApp,
+        isFlutterApp: () => Future.value(isFlutterApp),
         devtoolsLauncher: serveDevTools
             ? (hostname) async {
                 final server = await DevToolsServer().serveDevTools(
                   hostname: hostname,
                   enableStdinCommands: false,
-                  customDevToolsPath: sdkLayout.devToolsDirectory,
+                  customDevToolsPath: devToolsPath,
                 );
                 if (server == null) {
                   throw StateError('DevTools server could not be started.');