+++ dwds/test/readers/frontend_server_asset_reader_test.dart
@@ -3,16 +3,15 @@
 // BSD-style license that can be found in the LICENSE file.
 
 @Timeout(Duration(minutes: 2))
-import 'dart:async';
 import 'dart:io';
 
 import 'package:dwds/src/readers/frontend_server_asset_reader.dart';
 import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
-import 'package:test_common/test_sdk_layout.dart';
 
-import '../fixtures/project.dart';
 import '../fixtures/utilities.dart';
+
+final packagesDir = absolutePath(pathFromFixtures: '_test');
 
 final fixturesDir = absolutePath(pathFromDwds: p.join('test', 'fixtures'));
 
@@ -21,9 +20,6 @@ void main() {
   late Directory tempFixtures;
   late File jsonOriginal;
   late File mapOriginal;
-
-  final testProject = TestProject.testWithSoundNullSafety;
-  final packagesDir = testProject.absolutePackageDirectory;
 
   Future<void> createTempFixtures() async {
     tempFixtures = await Directory.systemTemp.createTemp('dwds_test_fixtures');
@@ -35,8 +31,7 @@ void main() {
   }
 
   setUpAll(() async {
-    final sdkLayout = TestSdkLayout.defaultSdkLayout;
-    await Process.run(sdkLayout.dartPath, ['pub', 'upgrade'],
+    await Process.run(dartPath, ['pub', 'upgrade'],
         workingDirectory: packagesDir);
   });
 
@@ -46,7 +41,7 @@ void main() {
       p.join(tempFixtures.path, 'main.dart.dill'),
       packagesDir,
     );
-    await assetReader.updateCaches();
+    assetReader.updateCaches();
   });
 
   tearDown(() async {
@@ -107,7 +102,7 @@ void main() {
             .writeAsString((await mapOriginal.readAsString())
                 .replaceAll('web/main.dart.lib.js', 'web/foo.dart.lib.js'));
 
-        await assetReader.updateCaches();
+        assetReader.updateCaches();
 
         final newResult =
             await assetReader.sourceMapContents('web/foo.dart.lib.js.map');