+++ test_common/lib/test_sdk_configuration.dart
@@ -2,10 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'dart:io';
-
+// TODO:(annagrin) Move to a test_common package.
 import 'package:dwds/sdk_configuration.dart';
-import 'package:logging/logging.dart';
 
 import 'package:test_common/sdk_asset_generator.dart';
 import 'package:test_common/test_sdk_layout.dart';
@@ -21,61 +19,26 @@ import 'package:test_common/test_sdk_layout.dart';
 /// TODO(annagrin): update to only generating missing sound artifacts
 /// for frontend server after we have no uses of weak null safety.
 class TestSdkConfigurationProvider extends SdkConfigurationProvider {
-  final _logger = Logger('TestSdkConfigurationProvider');
-
-  final bool _verbose;
-  late final Directory _sdkDirectory;
+  final bool _verboseCompiler;
   SdkConfiguration? _configuration;
 
-  late final TestSdkLayout sdkLayout;
+  final sdkLayout = TestSdkLayout.defaultSdkLayout;
 
-  TestSdkConfigurationProvider({bool verbose = false}) : _verbose = verbose {
-    _sdkDirectory = Directory.systemTemp.createTempSync('sdk copy');
-    sdkLayout = TestSdkLayout.createDefault(_sdkDirectory.path);
-  }
+  TestSdkConfigurationProvider({bool verboseCompiler = false})
+      : _verboseCompiler = verboseCompiler;
 
   @override
   Future<SdkConfiguration> get configuration async =>
       _configuration ??= await _create();
 
   /// Generate missing assets in the default SDK layout.
-  ///
-  /// Creates a copy of the SDK directory where all the missing assets
-  /// are generated. Tests using this configuration run using the copy
-  /// sdk layout to make sure the actual SDK is not modified.
   Future<SdkConfiguration> _create() async {
-    try {
-      await copyDirectory(
-          TestSdkLayout.defaultSdkDirectory, _sdkDirectory.path);
-    } catch (e, s) {
-      _logger.severe('Failed to create SDK directory copy', e, s);
-      dispose();
-      rethrow;
-    }
+    final assetGenerator = SdkAssetGenerator(
+      sdkLayout: sdkLayout,
+      verboseCompiler: _verboseCompiler,
+    );
 
-    try {
-      final assetGenerator = SdkAssetGenerator(
-        sdkLayout: sdkLayout,
-        verboseCompiler: _verbose,
-      );
-
-      await assetGenerator.generateSdkAssets();
-      return TestSdkLayout.createConfiguration(sdkLayout);
-    } catch (e, s) {
-      _logger.severe('Failed generate missing assets', e, s);
-      dispose();
-      rethrow;
-    }
-  }
-
-  void dispose({bool retry = true}) {
-    try {
-      if (_sdkDirectory.existsSync()) {
-        _sdkDirectory.deleteSync(recursive: true);
-      }
-    } catch (e, s) {
-      _logger.warning('Failed delete SDK directory copy', e, s);
-      dispose(retry: false);
-    }
+    await assetGenerator.generateSdkAssets();
+    return TestSdkLayout.defaultSdkConfiguration;
   }
 }