+++ dwds/CHANGELOG.md
@@ -11,10 +11,14 @@
   - Display records as a container of fields.
 - Remove test-only code from `sdk_configuration.dart`.
 - Move shared test-only code to a new `test_common` package.
+- Convert unnecessary async code to sync.
 
 **Breaking changes**
+
 - Require `sdkConfigurationProvider` in `ExpressionCompilerService`
   constructor.
+- Change DWDS parameter `isFlutterApp` from type `bool?` to type
+  `Future<bool> Function()?`.
 
 ## 17.0.0
 
@@ -40,6 +44,7 @@
 - Fix expression compiler throwing when weak SDK summary is not found.
 
 **Breaking changes**
+
 - Include an optional param to `Dwds.start` to indicate whether it is running
   internally or externally.
 - Include an optional param to `Dwds.start` to indicate whether it a Flutter